ref(tests): split out tests from the config test into their own smaller tests
