+++ rootfs/api/tests/test_config.py
@@ -21,7 +21,6 @@ import requests_mock
 @mock.patch('api.models.release.publish_release', lambda *args: None)
 @mock.patch('api.models.release.docker_get_port', mock_port)
 class ConfigTest(DeisTransactionTestCase):
-
     """Tests setting and updating config values"""
 
     fixtures = ['tests.json']
@@ -302,407 +301,6 @@ class ConfigTest(DeisTransactionTestCase):
         self.assertIn('PORT', response.data['values'])
         return response
 
-    def test_limit_memory(self, mock_requests):
-        """
-        Test that limit is auto-created for a new app and that
-        limits can be updated using a PATCH
-        """
-        app_id = self.create_app()
-        url = '/v2/apps/{app_id}/config'.format(**locals())
-
-        # check default limit
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('memory', response.data)
-        self.assertEqual(response.data['memory'], {})
-        # regression test for https://github.com/deis/deis/issues/1563
-        self.assertNotIn('"', response.data['memory'])
-
-        # set an initial limit
-        mem = {'web': '1G'}
-        body = {'memory': json.dumps(mem)}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit1 = response.data
-
-        # check memory limits
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('memory', response.data)
-        memory = response.data['memory']
-        self.assertIn('web', memory)
-        self.assertEqual(memory['web'], '1G')
-
-        # set an additional value
-        body = {'memory': json.dumps({'worker': '512M'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit2 = response.data
-        self.assertNotEqual(limit1['uuid'], limit2['uuid'])
-        memory = response.data['memory']
-        self.assertIn('worker', memory)
-        self.assertEqual(memory['worker'], '512M')
-        self.assertIn('web', memory)
-        self.assertEqual(memory['web'], '1G')
-
-        # read the limit again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        limit3 = response.data
-        self.assertEqual(limit2, limit3)
-        memory = response.data['memory']
-        self.assertIn('worker', memory)
-        self.assertEqual(memory['worker'], '512M')
-        self.assertIn('web', memory)
-        self.assertEqual(memory['web'], '1G')
-
-        # regression test for https://github.com/deis/deis/issues/1613
-        # ensure that config:set doesn't wipe out previous limits
-        body = {'values': json.dumps({'NEW_URL2': 'http://localhost:8080/'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        self.assertIn('NEW_URL2', response.data['values'])
-
-        # read the limit again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        memory = response.data['memory']
-        self.assertIn('worker', memory)
-        self.assertEqual(memory['worker'], '512M')
-        self.assertIn('web', memory)
-        self.assertEqual(memory['web'], '1G')
-
-        # unset a value
-        body = {'memory': json.dumps({'worker': None})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit4 = response.data
-        self.assertNotEqual(limit3['uuid'], limit4['uuid'])
-        self.assertNotIn('worker', json.dumps(response.data['memory']))
-
-        # bad memory values
-        mem = {'web': '1Z'}
-        body = {'memory': json.dumps(mem)}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        mem = {'w3&b': '1G'}
-        body = {'memory': json.dumps(mem)}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # disallow put/patch/delete
-        response = self.client.put(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.patch(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.delete(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        return limit4
-
-    def test_limit_cpu(self, mock_requests):
-        """
-        Test that CPU limits can be set
-        """
-        app_id = self.create_app()
-        url = '/v2/apps/{app_id}/config'.format(**locals())
-
-        # check default limit
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('cpu', response.data)
-        self.assertEqual(response.data['cpu'], {})
-        # regression test for https://github.com/deis/deis/issues/1563
-        self.assertNotIn('"', response.data['cpu'])
-
-        # set an initial limit
-        body = {'cpu': json.dumps({'web': '1024'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit1 = response.data
-
-        # check memory limits
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('cpu', response.data)
-        cpu = response.data['cpu']
-        self.assertIn('web', cpu)
-        self.assertEqual(cpu['web'], '1024')
-
-        # set an additional value
-        body = {'cpu': json.dumps({'worker': '512m'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit2 = response.data
-        self.assertNotEqual(limit1['uuid'], limit2['uuid'])
-        cpu = response.data['cpu']
-        self.assertIn('worker', cpu)
-        self.assertEqual(cpu['worker'], '512m')
-        self.assertIn('web', cpu)
-        self.assertEqual(cpu['web'], '1024')
-
-        # read the limit again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        limit3 = response.data
-        self.assertEqual(limit2, limit3)
-        cpu = response.data['cpu']
-        self.assertIn('worker', cpu)
-        self.assertEqual(cpu['worker'], '512m')
-        self.assertIn('web', cpu)
-        self.assertEqual(cpu['web'], '1024')
-
-        # unset a value
-        body = {'cpu': json.dumps({'worker': None})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        limit4 = response.data
-        self.assertNotEqual(limit3['uuid'], limit4['uuid'])
-        self.assertNotIn('worker', json.dumps(response.data['cpu']))
-
-        # bad cpu values
-        mem = {'web': '1G'}
-        body = {'cpu': json.dumps(mem)}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        mem = {'w3&b': '1G'}
-        body = {'cpu': json.dumps(mem)}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # disallow put/patch/delete
-        response = self.client.put(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.patch(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.delete(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        return limit4
-
-    def test_tags(self, mock_requests):
-        """
-        Test that tags can be set on an application
-        """
-        app_id = self.create_app()
-
-        # check default
-        url = '/v2/apps/{app_id}/config'.format(**locals())
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('tags', response.data)
-        self.assertEqual(response.data['tags'], {})
-
-        # set some tags
-        body = {'tags': json.dumps({'environ': 'dev'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        tags1 = response.data
-
-        # check tags again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('tags', response.data)
-        tags = response.data['tags']
-        self.assertIn('environ', tags)
-        self.assertEqual(tags['environ'], 'dev')
-
-        # set an additional value
-        body = {'tags': json.dumps({'rack': '1'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        tags2 = response.data
-        self.assertNotEqual(tags1['uuid'], tags2['uuid'])
-        tags = response.data['tags']
-        self.assertIn('rack', tags)
-        self.assertEqual(tags['rack'], '1')
-        self.assertIn('environ', tags)
-        self.assertEqual(tags['environ'], 'dev')
-
-        # read the limit again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        tags3 = response.data
-        self.assertEqual(tags2, tags3)
-        tags = response.data['tags']
-        self.assertIn('rack', tags)
-        self.assertEqual(tags['rack'], '1')
-        self.assertIn('environ', tags)
-        self.assertEqual(tags['environ'], 'dev')
-
-        # unset a value
-        body = {'tags': json.dumps({'rack': None})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        tags4 = response.data
-        self.assertNotEqual(tags3['uuid'], tags4['uuid'])
-        self.assertNotIn('rack', json.dumps(response.data['tags']))
-
-        # set valid values
-        body = {'tags': json.dumps({'kubernetes.io/hostname': '172.17.8.100'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        body = {'tags': json.dumps({'is.valid': 'is-also_valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        body = {'tags': json.dumps({'host.the-name.com/is.valid': 'valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        body = {'tags': json.dumps({'host.the-name.com/does.no.exist': 'valid'})}
-        response = self.client.post(url, body)
-        self.assertContains(
-            response,
-            'Addition of host.the-name.com/does.no.exist=valid is the cause',
-            status_code=400
-        )
-
-        # set invalid values
-        body = {'tags': json.dumps({'valid': 'in\nvalid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        body = {'tags': json.dumps({'host.name.com/notvalid-': 'valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        body = {'tags': json.dumps({'valid': 'invalid.'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        body = {'tags': json.dumps({'host.name.com/,not.valid': 'valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        long_tag = 'a' * 300
-        body = {'tags': json.dumps({'{}/not.valid'.format(long_tag): 'valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        body = {'tags': json.dumps({'this&foo.com/not.valid': 'valid'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # disallow put/patch/delete
-        response = self.client.put(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.patch(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.delete(url)
-        self.assertEqual(response.status_code, 405, response.data)
-
-    def test_registry(self, mock_requests):
-        """
-        Test that registry information can be set on an application
-        """
-        app_id = self.create_app()
-
-        # check default
-        url = '/v2/apps/{app_id}/config'.format(**locals())
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('registry', response.data)
-        self.assertEqual(response.data['registry'], {})
-
-        # set some registry information without PORT
-        body = {'registry': json.dumps({'username': 'bob'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-        registry1 = response.data
-
-        # set required PORT
-        body = {'values': json.dumps({'PORT': '80'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        registry1 = response.data
-
-        # set some registry information
-        body = {'registry': json.dumps({'username': 'bob'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        registry1 = response.data
-
-        # check registry information again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        self.assertIn('registry', response.data)
-        registry = response.data['registry']
-        self.assertIn('username', registry)
-        self.assertEqual(registry['username'], 'bob')
-
-        # set an additional value
-        # set them upper case, internally it should translate to lower
-        body = {'registry': json.dumps({'PASSWORD': 's3cur3pw1'})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        registry2 = response.data
-        self.assertNotEqual(registry1['uuid'], registry2['uuid'])
-        registry = response.data['registry']
-        self.assertIn('password', registry)
-        self.assertEqual(registry['password'], 's3cur3pw1')
-        self.assertIn('username', registry)
-        self.assertEqual(registry['username'], 'bob')
-
-        # read the registry information again
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200, response.data)
-        registry3 = response.data
-        self.assertEqual(registry2, registry3)
-        registry = response.data['registry']
-        self.assertIn('password', registry)
-        self.assertEqual(registry['password'], 's3cur3pw1')
-        self.assertIn('username', registry)
-        self.assertEqual(registry['username'], 'bob')
-
-        # unset a value
-        body = {'registry': json.dumps({'password': None})}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-        registry4 = response.data
-        self.assertNotEqual(registry3['uuid'], registry4['uuid'])
-        self.assertNotIn('password', json.dumps(response.data['registry']))
-
-        # bad registry key values
-        body = {'registry': json.dumps({'pa$$w0rd': 'woop'})}
-        response = self.client.post(url, body)
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # disallow put/patch/delete
-        response = self.client.put(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.patch(url)
-        self.assertEqual(response.status_code, 405, response.data)
-        response = self.client.delete(url)
-        self.assertEqual(response.status_code, 405, response.data)
-
-    def test_registry_deploy(self, mock_requests):
-        """
-        Test that registry information can be applied
-        """
-        app_id = self.create_app()
-
-        # Set mandatory PORT
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'PORT': '4999'})}
-        )
-
-        # Set registry information
-        body = {'registry': json.dumps({
-            'username': 'bob',
-            'password': 's3cur3pw1'
-        })}
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            body
-        )
-        self.assertEqual(response.status_code, 201, response.data)
-        self.assertIn('username', response.data['registry'])
-        self.assertIn('password', response.data['registry'])
-        self.assertEqual(response.data['registry']['username'], 'bob')
-        self.assertEqual(response.data['registry']['password'], 's3cur3pw1')
-
-        # post a new build
-        url = "/v2/apps/{app_id}/builds".format(**locals())
-        body = {'image': 'autotest/example'}
-        response = self.client.post(url, body)
-        self.assertEqual(response.status_code, 201, response.data)
-
     def test_config_owner_is_requesting_user(self, mock_requests):
         """
         Ensure that setting the config value is owned by the requesting user
@@ -727,188 +325,3 @@ class ConfigTest(DeisTransactionTestCase):
         body = {'values': {'FOO': 'bar'}}
         response = self.client.post(url, body)
         self.assertEqual(response.status_code, 403)
-
-    def test_healthchecks_validations(self, mock_requests):
-        """
-        Test that healthchecks validations work
-        """
-        app_id = self.create_app()
-
-        # Set one of the values that require a numeric value to a string
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_INITIAL_DELAY': 'horse'})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # test URL - Path is the only allowed thing
-        # Try setting various things such as query param
-
-        # query param
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_URL': '/health?testing=0'})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # fragment
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_URL': '/health#db'})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # netloc
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_URL': 'http://someurl.com/health/'})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # no path
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_URL': 'http://someurl.com'})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-    def test_config_healthchecks(self, mock_requests):
-        """
-        Test that healthchecks can be applied
-        """
-        app_id = self.create_app()
-        readiness_probe = {'healthcheck': {'readinessProbe': {'httpGet': {'port': 5000}}}}
-
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            readiness_probe)
-        self.assertEqual(response.status_code, 201, response.data)
-        self.assertIn('readinessProbe', response.data['healthcheck'])
-        self.assertEqual(response.data['healthcheck'], readiness_probe['healthcheck'])
-
-        liveness_probe = {'healthcheck': {'livenessProbe':
-                                          {'httpGet': {'port': 5000},
-                                           'successThreshold': 1}}}
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            liveness_probe)
-        self.assertEqual(response.status_code, 201, response.data)
-        self.assertIn('livenessProbe', response.data['healthcheck'])
-        self.assertEqual(
-            response.data['healthcheck']['livenessProbe'],
-            liveness_probe['healthcheck']['livenessProbe'])
-        # check that the readiness probe is still there too!
-        self.assertIn('readinessProbe', response.data['healthcheck'])
-        self.assertEqual(
-            response.data['healthcheck']['readinessProbe'],
-            readiness_probe['healthcheck']['readinessProbe'])
-
-        # post a new build
-        response = self.client.post(
-            "/v2/apps/{app_id}/builds".format(**locals()),
-            {'image': 'quay.io/autotest/example'}
-        )
-        self.assertEqual(response.status_code, 201, response.data)
-
-    def test_config_healthchecks_validations(self, mock_requests):
-        """
-        Test that healthchecks validations work
-        """
-        app_id = self.create_app()
-
-        # Set one of the values that require a numeric value to a string
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'healthcheck': json.dumps({'livenessProbe':
-                                        {'httpGet': {'port': '50'}, 'initialDelaySeconds': "t"}})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # Don't set one of the mandatory value
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'healthcheck': json.dumps({'livenessProbe':
-                                        {'httpGet': {'path': '/'}, 'initialDelaySeconds': 1}})}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-        # set liveness success threshold to a non-1 value
-        # Don't set one of the mandatory value
-        response = self.client.post(
-            '/v2/apps/{app_id}/config'.format(**locals()),
-            {'healthcheck': {'livenessProbe':
-                             {'httpGet': {'path': '/', 'port': 5000},
-                              'successThreshold': 5}}}
-        )
-        self.assertEqual(response.status_code, 400, response.data)
-
-    def test_config_healthchecks_legacy(self, mock_requests):
-        """
-        Test that when a user uses `deis config:set HEALTHCHECK_URL=/`, the config
-        object is rolled over to the `healthcheck` field.
-        """
-        app_id = self.create_app()
-        app = App.objects.get(id=app_id)
-
-        # Set healthcheck URL to get defaults set
-        response = self.client.post(
-            '/v2/apps/{app.id}/config'.format(**locals()),
-            {'values': json.dumps({'HEALTHCHECK_URL': '/health'})}
-        )
-        self.assertEqual(response.status_code, 201, response.data)
-        # this gets migrated to the new healtcheck format
-        self.assertNotIn('HEALTHCHECK_URL', response.data['values'])
-        # legacy defaults
-        expected = {
-            'livenessProbe': {
-                'initialDelaySeconds': 50,
-                'timeoutSeconds': 50,
-                'periodSeconds': 10,
-                'successThreshold': 1,
-                'failureThreshold': 3,
-                'httpGet': {
-                    'path': '/health'
-                }
-            },
-            'readinessProbe': {
-                'initialDelaySeconds': 50,
-                'timeoutSeconds': 50,
-                'periodSeconds': 10,
-                'successThreshold': 1,
-                'failureThreshold': 3,
-                'httpGet': {
-                    'path': '/health'
-                }
-            }
-        }
-        actual = app.config_set.latest().healthcheck
-        self.assertEqual(actual, expected)
-        # Now set all the envvars and check to make sure they are written properly
-        response = self.client.post(
-            '/v2/apps/{app.id}/config'.format(**locals()),
-            {
-                'values': json.dumps({
-                    'HEALTHCHECK_URL': '/health',
-                    'HEALTHCHECK_INITIAL_DELAY': '25',
-                    'HEALTHCHECK_TIMEOUT': '10',
-                    'HEALTHCHECK_PERIOD_SECONDS': '5',
-                    'HEALTHCHECK_SUCCESS_THRESHOLD': '2',
-                    'HEALTHCHECK_FAILURE_THRESHOLD': '2'})
-            }
-        )
-        self.assertEqual(response.status_code, 201, response.data)
-        # this gets migrated to the new healtcheck format
-        self.assertNotIn('HEALTHCHECK_INITIAL_DELAY', response.data['values'])
-        expected['livenessProbe'] = {
-            'initialDelaySeconds': 25,
-            'timeoutSeconds': 10,
-            'periodSeconds': 5,
-            'successThreshold': 2,
-            'failureThreshold': 2,
-            'httpGet': {
-                'path': '/health'
-            }
-        }
-        expected['readinessProbe'] = expected['livenessProbe']
-        actual = app.config_set.latest().healthcheck
-        self.assertEqual(expected, actual)
